feat(CRUD):  creating the logic for the CRUD of logistics operators and documenting their possible responses

+++ backend/src/logistic-operator/logistic-operator.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  Injectable,
+  InternalServerErrorException,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateLogisticOperatorDto } from './dto/create-logistic-operator.dto';
 import { UpdateLogisticOperatorDto } from './dto/update-logistic-operator.dto';
 import { LogisticOperatorRepository } from './logistic-operator.repository';
@@ -7,23 +11,68 @@ import { LogisticOperatorRepository } from './logistic-operator.repository';
 export class LogisticOperatorService {
   constructor(private readonly repository: LogisticOperatorRepository) {}
 
-  create(createLogisticOperatorDto: CreateLogisticOperatorDto) {
-    return this.repository.create(createLogisticOperatorDto);
+  async create(createLogisticOperatorDto: CreateLogisticOperatorDto) {
+    try {
+      return await this.repository.create(createLogisticOperatorDto);
+    } catch (error) {
+      throw new InternalServerErrorException(
+        'Error creating Logistic Operator',
+      );
+    }
   }
 
-  findAll() {
-    return this.repository.findMany();
+  async findAll() {
+    try {
+      return await this.repository.findMany();
+    } catch (error) {
+      throw new InternalServerErrorException(
+        'Error fetching Logistic Operators',
+      );
+    }
   }
 
-  findOne(id: string) {
-    return this.repository.findOne(id);
+  async findOne(id: string) {
+    const logisticOperator = await this.repository.findOne(id);
+    if (!logisticOperator) {
+      throw new NotFoundException('Logistic Operator not found');
+    }
+    try {
+      return logisticOperator;
+    } catch (error) {
+      throw new InternalServerErrorException(
+        'Error fetching Logistic Operator',
+      );
+    }
   }
 
-  update(id: string, updateLogisticOperatorDto: UpdateLogisticOperatorDto) {
-    return this.repository.update(id, updateLogisticOperatorDto);
+  async update(
+    id: string,
+    updateLogisticOperatorDto: UpdateLogisticOperatorDto,
+  ) {
+    const logisticOperator = await this.findOne(id);
+    if (!logisticOperator) {
+      throw new NotFoundException('Logistic Operator not found');
+    }
+    try {
+      return await this.repository.update(id, updateLogisticOperatorDto);
+    } catch (error) {
+      throw new InternalServerErrorException(
+        'Error updating Logistic Operator',
+      );
+    }
   }
 
-  remove(id: string) {
-    return this.repository.remove(id);
+  async remove(id: string) {
+    const logisticOperator = await this.findOne(id);
+    if (!logisticOperator) {
+      throw new NotFoundException('Logistic Operator not found');
+    }
+    try {
+      return await this.repository.remove(id);
+    } catch (error) {
+      throw new InternalServerErrorException(
+        'Error deleting Logistic Operator',
+      );
+    }
   }
 }
